pyproject.toml
[project]
name = "a2a-distributed-tracing"
version = "0.1.0"
description = "Agent-to-Agent conversation simulation using LangGraph A2A protocol"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.0.3",
    "langchain>=1.0.0",
    "langchain-openai>=1.0.2",
    "langgraph-cli[inmem]>=0.4.7",
    "python-dotenv>=1.2.1",
    "typing_extensions>=4.15.0",
    "aiohttp>=3.13.2",
    "google-adk[a2a]>=0.1.0",
    "uvicorn>=0.30.0",
    "litellm>=1.80.5",
]

